gemini-document-qa/requirements.txt
gradio
python-dotenv
google-genai
